Fix attempt 1

+++ src/nativeMain/kotlin/rinha/models/PaymentsSummary.kt
@@ -10,6 +10,6 @@ data class PaymentsSummary(
 
 @Serializable
 data class ProcessorSummary(
-    val totalRequests: Int,
+    val totalRequests: Long,
     val totalAmount: Double
 )
